Project corrections after milestone 2

+++ carPark-service/src/main/java/cz/muni/fi/pa165/service/CarServiceImpl.java
@@ -1,49 +1,79 @@
-/*
- * To change this template, choose Tools | Templates
- * and open the template in the editor.
- */
 package cz.muni.fi.pa165.service;
 
 import java.util.List;
 
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import cz.muni.fi.pa165.daos.CarDao;
+import cz.muni.fi.pa165.daos.CarDaoImpl;
 import cz.muni.fi.pa165.entities.Car;
 import cz.muni.fi.pa165.enums.Fuel;
 import cz.muni.fi.pa165.enums.Transmission;
-import cz.muni.fi.pa165.exceptions.CarParkServiceException;
+import java.util.ArrayList;
+import javax.inject.Inject;
+import org.springframework.context.annotation.ComponentScan;
 
 /**
  *
  * @author xcmarko
  */
 @Service
-public class CarServiceImpl implements CarService{
+@ComponentScan(basePackageClasses={CarDaoImpl.class})
+public class CarServiceImpl implements CarService {
     
-    @Autowired
+    @Inject
     private CarDao carDao;
 
     @Override
-    public Long createCar(Car car) {
-        Long carId = null;
-        try {
-            carId = carDao.createCar(car);
-        } catch (IllegalArgumentException | SecurityException e) {
-            throw new CarParkServiceException(e);
+    public Car createCar(Car car) {
+        if(carDao.createCar(car)) {
+            return car;
         }
-        
-        return carId;
+        return null;
     }
 
     @Override
-    public void removeCar(Car car) {
-        try {
-            carDao.deleteCar(car);
-        } catch (IllegalArgumentException | SecurityException e) {
-            throw new CarParkServiceException(e);
+    public boolean deleteCar(Car car) {
+        if (car != null) {
+            return carDao.deleteCar(car);
         }
+        return false;
+    }
+
+    @Override
+    public Car updateName(Car car, String newName) {
+        if(car != null && newName != null && !newName.isEmpty()) {
+            car.setName(newName);
+            return carDao.updateCar(car);
+        }
+        return null;
+    }
+
+    @Override
+    public Car updateColor(Car car, String newColor) {
+        if(car != null && newColor != null && !newColor.isEmpty()) {
+            car.setColor(newColor);
+            return carDao.updateCar(car);
+        }
+        return null;
+    }
+
+    @Override
+    public Car updateFuel(Car car, Fuel newFuel) {
+        if(car != null && newFuel != null) {
+            car.setFuel(newFuel);
+            return carDao.updateCar(car);
+        }
+        return null;
+    }
+
+    @Override
+    public Car updateTransmission(Car car, Transmission newTransmission) {
+        if(car != null && newTransmission != null) {
+            car.setTransmission(newTransmission);
+            return carDao.updateCar(car);
+        }
+        return null;
     }
 
     @Override
@@ -64,6 +94,23 @@ public class CarServiceImpl implements CarService{
     @Override
     public List<Car> getCarsByTransmission(Transmission transmission) {
         return carDao.findCarByTransmission(transmission);
+    }
+
+    @Override
+    public List<Car> getCarsByTransmissionAndFuel(Transmission transmission, Fuel fuel) {
+        List<Car> result = new ArrayList<>();
+        List<Car> allCarsByTransmission = this.getCarsByTransmission(transmission);
+
+        if(allCarsByTransmission.isEmpty()) {
+            return result;
+        }
+
+        for(Car item : allCarsByTransmission) {
+            if(item.getFuel() == fuel) {
+                result.add(item);
+            }
+        }
+        return result;
     }
     
 }
